# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BookLib API",
    "version": "1.0.0",
    "description": "API for managing books, users, tags, comments, ratings, and plugins. Tags are always ensured from a resource file."
  },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/books": {
      "get": {
        "tags": ["Books"],
        "summary": "List books",
        "responses": {"200": {"description": "List of books"}}
      },
      "post": {
        "tags": ["Books"],
        "summary": "Add book",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "isbn": {"type": "string"},
                  "description": {"type": "string"},
                  "publish_year": {"type": ["integer", "string", "null"]},
                  "series": {"type": ["string", "null"]},
                  "tags": {"type": "array", "items": {"type": "string"}},
                  "plugin": {"type": "string", "description": "Plugin to use for enrichment (GoogleBooksPlugin, OpenLibraryPlugin)"}
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {"201": {"description": "Book added"}},
        "security": [{"BearerAuth": []}]
      }
    },
    "/books/{id}": {
      "get": {
        "tags": ["Books"],
        "summary": "Get book",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Book details"}}
      },
      "put": {
        "tags": ["Books"],
        "summary": "Update book",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {"type": "string", "description": "Book title (auto-filled if ISBN provided)"},
                    "isbn": {"type": "string", "description": "ISBN to fetch book info from plugin"},
                    "publish_year": {"type": ["integer", "string", "null"], "description": "Publish year (auto-filled if ISBN provided)"},
                    "series": {"type": ["string", "null"], "description": "Book series (auto-filled if ISBN provided)"},
                    "tags": {"type": "array", "items": {"type": "string"}, "description": "Genres/subjects (auto-filled if ISBN provided)"},
                    "plugin": {"type": "string", "description": "Plugin to use for enrichment (GoogleBooksPlugin, OpenLibraryPlugin)"}
                  },
                  "required": ["title"]
                }
              }
            }
          },
          "responses": {
            "200": {"description": "Book updated"},
            "400": {"description": "Missing fields or invalid data"}
          },
        "security": [{"BearerAuth": []}]
      },
      "delete": {
        "tags": ["Books"],
        "summary": "Delete book",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Book deleted"}},
        "security": [{"BearerAuth": []}]
      }
    },
    "/books/{id}/recheck": {
      "post": {
        "tags": ["Books"],
        "summary": "Recheck and update book info",
        "description": "Rechecks and updates book info by ID using the selected plugin (default: GoogleBooksPlugin).",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plugin": {"type": "string", "description": "Plugin to use for enrichment (GoogleBooksPlugin, OpenLibraryPlugin)"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Book info rechecked and updated"},
          "400": {"description": "Error or plugin not found"}
        },
        "security": [{"BearerAuth": []}]
      }
    },
    "/tags": {
      "get": {
        "tags": ["Tags"],
        "summary": "Get all tags",
        "description": "Returns all tags in the database. Required tags are always loaded from a resource file.",
        "responses": {"200": {"description": "List of tags", "content": {"application/json": {"schema": {"type": "array", "items": {"type": "object", "properties": {"id": {"type": "integer"}, "name": {"type": "string"}}}}}}}}
      },
      "post": {
        "tags": ["Tags"],
        "summary": "Add a new tag",
        "description": "Adds a new tag to the database.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"name": {"type": "string"}}, "required": ["name"]}}}},
        "responses": {"201": {"description": "Tag added", "content": {"application/json": {"schema": {"type": "object", "properties": {"id": {"type": "integer"}, "name": {"type": "string"}}}}}}},
        "security": [{"BearerAuth": []}]
      }
    },
...
    "/plugins/load": {"post": {"summary": "Load plugin", "requestBody": {"required": true}, "responses": {"200": {"description": "Plugin loaded"}}}},
    "/plugins/unload": {"post": {"summary": "Unload plugin", "requestBody": {"required": true}, "responses": {"200": {"description": "Plugin unloaded"}}}},
      "/plugins/{plugin_name}/run": {
        "post": {
          "tags": ["Plugins"],
          "summary": "Run a plugin",
          "description": "Executes the specified plugin with provided input data.",
          "parameters": [
            {"name": "plugin_name", "in": "path", "required": true, "schema": {"type": "string"}}
          ],
          "requestBody": {
            "required": false,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Input data for the plugin (e.g., ISBN for Goodreads)."
                }
              }
            }
          },
          "responses": {
            "200": {"description": "Plugin executed successfully"},
            "404": {"description": "Plugin not found"},
            "400": {"description": "Bad request or plugin error"}
          }
        }
      },
    "/books/{id}/full": {
      "get": {
        "tags": ["Books"],
        "summary": "Get book info, ratings, and comments",
        "description": "Returns book information, average rating, all ratings, and all comments for the specified book.",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {
            "description": "Book info, ratings, and comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "integer"},
                    "title": {"type": "string"},
                    "author": {"type": "string"},
                    "description": {"type": "string"},
                    "publish_year": {"type": ["integer", "null"]},
                    "series": {"type": ["string", "null"]},
                    "average_rating": {"type": ["number", "null"]},
                    "ratings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {"type": "integer"},
                          "user_id": {"type": "integer"},
                          "rating": {"type": "integer"}
                        }
                      }
                    },
                    "comments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {"type": "integer"},
                          "user_id": {"type": "integer"},
                          "content": {"type": "string"}
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {"description": "Book not found"}
        }
      }
    }
  }
}